tests/wpt/tests/css/css-flexbox/parsing/flex-flow-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout Module Level 1: parsing flex-flow with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-flow">
<meta name="assert" content="flex-flow supports only the grammar '<flex-direction> || <flex-wrap>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flex-flow", "nowrap row nowrap");
test_invalid_value("flex-flow", "column wrap column");
</script>
</body>
</html>
